Rename types to gtypes
