+++ guppy/compiler/core.py
@@ -4,7 +4,7 @@ from typing import Optional, Iterator
 
 from guppy.ast_util import AstNode
 from guppy.checker.core import Variable, CallableVariable
-from guppy.types import FunctionType
+from guppy.gtypes import FunctionType
 from guppy.hugr.hugr import OutPortV, DFContainingNode, Hugr
