+++ guppy/gtypes.py
